+++ packages/contracts/contracts/test/unit/TellerAndHandler.t.sol
@@ -31,6 +31,7 @@ import {SimpleERC1155Token} from "../tokens/SimpleERC1155Token.sol";
 import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import {Utils} from "../../libs/Utils.sol";
 import {AssetUtils} from "../../libs/AssetUtils.sol";
+import {Validation} from "../../libs/Validation.sol";
 import "../../libs/Types.sol";
 
 contract TellerAndHandlerTest is Test, ForgeUtils, PoseidonDeployer {
@@ -453,6 +454,57 @@ contract TellerAndHandlerTest is Test, ForgeUtils, PoseidonDeployer {
         teller.depositFunds(deposit);
     }
 
+    function testCompleteDepositFailsInvalidFields() public {
+        // Allow ALICE to direct deposit to teller
+        teller.setDepositSourcePermission(ALICE, true);
+
+        // Deploy and dep manager whitelist new token but not in handler
+        SimpleERC20Token token = ERC20s[0];
+        token.reserveTokens(
+            ALICE,
+            Validation.NOCTURNE_MAX_NOTE_VALUE + PER_NOTE_AMOUNT
+        );
+
+        // Approve 50M tokens for deposit
+        vm.prank(ALICE);
+        token.approve(
+            address(teller),
+            Validation.NOCTURNE_MAX_NOTE_VALUE + PER_NOTE_AMOUNT
+        );
+
+        // Valid deposit works
+        Deposit memory deposit = NocturneUtils.formatDeposit(
+            ALICE,
+            address(token),
+            PER_NOTE_AMOUNT,
+            NocturneUtils.ERC20_ID,
+            NocturneUtils.defaultStealthAddress()
+        );
+        vm.prank(ALICE);
+        teller.depositFunds(deposit);
+
+        // value > 2^252
+        Deposit memory badValueDeposit = deposit;
+        badValueDeposit.value = Validation.NOCTURNE_MAX_NOTE_VALUE + 1;
+        vm.prank(ALICE);
+        vm.expectRevert("invalid note");
+        teller.depositFunds(badValueDeposit);
+
+        // asset id > field modulus
+        Deposit memory badAssetIdDeposit = deposit;
+        badAssetIdDeposit.encodedAsset.encodedAssetId = type(uint256).max;
+        vm.prank(ALICE);
+        vm.expectRevert("invalid note");
+        teller.depositFunds(badAssetIdDeposit);
+
+        // asset addr > field modulus
+        Deposit memory badAssetAddrDeposit = deposit;
+        badAssetAddrDeposit.encodedAsset.encodedAssetAddr = type(uint256).max;
+        vm.prank(ALICE);
+        vm.expectRevert("Invalid encodedAssetAddr");
+        teller.depositFunds(badAssetAddrDeposit);
+    }
+
     function testProcessBundleTransferSingleJoinSplitWithBundlerComp() public {
         // Alice starts with 50M tokens in teller
         SimpleERC20Token token = ERC20s[0];
@@ -670,6 +722,61 @@ contract TellerAndHandlerTest is Test, ForgeUtils, PoseidonDeployer {
         assertEq(token.balanceOf(address(handler)), uint256(1));
         assertEq(token.balanceOf(address(ALICE)), uint256(0));
         assertEq(token.balanceOf(address(BOB)), uint256(4 * PER_NOTE_AMOUNT));
+    }
+
+    function testProcessBundleFailureInvalidRefundAddr() public {
+        // Alice starts with 2 * 50M in teller
+        SimpleERC20Token token = ERC20s[0];
+        reserveAndDepositFunds(ALICE, token, 2 * PER_NOTE_AMOUNT);
+
+        TrackedAsset[] memory trackedRefundAssets = new TrackedAsset[](0);
+
+        // Create operation with invalid refund addr
+        Bundle memory bundle = Bundle({operations: new Operation[](1)});
+        bundle.operations[0] = NocturneUtils.formatOperation(
+            FormatOperationArgs({
+                joinSplitTokens: NocturneUtils._joinSplitTokensArrayOfOneToken(
+                    address(token)
+                ),
+                joinSplitRefundValues: new uint256[](1),
+                gasToken: address(token),
+                root: handler.root(),
+                joinSplitsPublicSpends: NocturneUtils
+                    ._publicSpendsArrayOfOnePublicSpendArray(
+                        NocturneUtils.fillJoinSplitPublicSpends(
+                            PER_NOTE_AMOUNT,
+                            1
+                        )
+                    ),
+                trackedRefundAssets: trackedRefundAssets,
+                gasAssetRefundThreshold: 0,
+                executionGasLimit: DEFAULT_GAS_LIMIT,
+                gasPrice: 50,
+                actions: NocturneUtils.formatSingleTransferActionArray(
+                    address(token),
+                    BOB,
+                    PER_NOTE_AMOUNT
+                ),
+                atomicActions: false,
+                operationFailureType: OperationFailureType.NONE
+            })
+        );
+
+        bundle.operations[0].refundAddr = CompressedStealthAddress(0, 0);
+
+        vmExpectOperationProcessed(
+            ExpectOperationProcessedArgs({
+                maybeFailureReason: "invalid point",
+                assetsUnwrapped: false
+            })
+        );
+        OperationResult[] memory opResults = teller.processBundle(bundle);
+
+        // One op, processed = false, unwrapped = false
+        assertEq(opResults.length, uint256(1));
+        assertEq(opResults[0].opProcessed, false);
+        assertEq(opResults[0].assetsUnwrapped, false);
+        assertEq(opResults[0].failureReason, "invalid point");
     }
 
     function testProcessBundleFailureBadRoot() public {