+++ packages/contracts/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 
-- add `validateNote` and `validateCompressedPoint` to `Utils`
+- add `validateNote` and `validateCompressedPoint` to new `Validation` lib
 - validate refund notes before they are queued for insertion into the commitment tree
 - add test lib `AlgebraicUtils` that implements BN254 field arithmetic and point decrompression
 - append `bitmap` to preimage of `accumulatorHash` to prevent prover from lying about insertion type (TOB-4)